Extract shell mutation helpers
