+++ clients/apple/AlanNative/ShellModel.swift
@@ -167,1235 +167,3 @@ func shellPaneTitleBarTitle(for pane: ShellPane) -> String {
 
     return "Terminal"
 }
-
-extension ShellPaneTreeNode {
-    func resizingSplit(
-        _ targetNodeID: String,
-        ratio targetRatio: Double
-    ) -> (node: ShellPaneTreeNode, changed: Bool) {
-        if kind == .split, nodeID == targetNodeID {
-            return (
-                ShellPaneTreeNode(
-                    nodeID: nodeID,
-                    kind: kind,
-                    direction: direction,
-                    ratio: targetRatio,
-                    paneID: paneID,
-                    children: children
-                ),
-                true
-            )
-        }
-
-        guard let children, !children.isEmpty else { return (self, false) }
-        var didChange = false
-        let nextChildren = children.map { child in
-            let result = child.resizingSplit(targetNodeID, ratio: targetRatio)
-            didChange = didChange || result.changed
-            return result.node
-        }
-
-        guard didChange else { return (self, false) }
-        return (
-            ShellPaneTreeNode(
-                nodeID: nodeID,
-                kind: kind,
-                direction: direction,
-                ratio: ratio,
-                paneID: paneID,
-                children: nextChildren
-            ),
-            true
-        )
-    }
-
-    func equalizedSplits() -> ShellPaneTreeNode {
-        switch kind {
-        case .pane:
-            return self
-        case .split:
-            return ShellPaneTreeNode(
-                nodeID: nodeID,
-                kind: kind,
-                direction: direction,
-                ratio: 0.5,
-                paneID: paneID,
-                children: children?.map { $0.equalizedSplits() }
-            )
-        }
-    }
-
-    func splittingPane(
-        _ targetPaneID: String,
-        direction: ShellSplitDirection,
-        splitNodeID: String,
-        newLeafNodeID: String,
-        newPaneID: String
-    ) -> ShellPaneTreeNode {
-        splittingPane(
-            targetPaneID,
-            placement: .defaultPlacement(for: direction),
-            splitNodeID: splitNodeID,
-            newLeafNodeID: newLeafNodeID,
-            newPaneID: newPaneID
-        )
-    }
-
-    func splittingPane(
-        _ targetPaneID: String,
-        placement: ShellPaneSplitDirection,
-        splitNodeID: String,
-        newLeafNodeID: String,
-        newPaneID: String
-    ) -> ShellPaneTreeNode {
-        switch kind {
-        case .pane:
-            guard paneID == targetPaneID else { return self }
-            let currentLeaf = ShellPaneTreeNode(
-                nodeID: nodeID,
-                kind: .pane,
-                direction: nil,
-                paneID: targetPaneID,
-                children: nil
-            )
-            let newLeaf = ShellPaneTreeNode(
-                nodeID: newLeafNodeID,
-                kind: .pane,
-                direction: nil,
-                paneID: newPaneID,
-                children: nil
-            )
-            return ShellPaneTreeNode(
-                nodeID: splitNodeID,
-                kind: .split,
-                direction: placement.splitDirection,
-                ratio: 0.5,
-                paneID: nil,
-                children: placement.placesNewPaneBeforeTarget
-                    ? [newLeaf, currentLeaf]
-                    : [currentLeaf, newLeaf]
-            )
-        case .split:
-            return ShellPaneTreeNode(
-                nodeID: nodeID,
-                kind: .split,
-                direction: self.direction,
-                ratio: ratio,
-                paneID: nil,
-                children: (children ?? []).map {
-                    $0.splittingPane(
-                        targetPaneID,
-                        placement: placement,
-                        splitNodeID: splitNodeID,
-                        newLeafNodeID: newLeafNodeID,
-                        newPaneID: newPaneID
-                    )
-                }
-            )
-        }
-    }
-
-    func removingPane(_ targetPaneID: String) -> ShellPaneTreeNode? {
-        switch kind {
-        case .pane:
-            return paneID == targetPaneID ? nil : self
-        case .split:
-            let remainingChildren = (children ?? []).compactMap { $0.removingPane(targetPaneID) }
-            if remainingChildren.isEmpty {
-                return nil
-            }
-            if remainingChildren.count == 1 {
-                return remainingChildren[0]
-            }
-            return ShellPaneTreeNode(
-                nodeID: nodeID,
-                kind: .split,
-                direction: direction,
-                ratio: ratio,
-                paneID: nil,
-                children: remainingChildren
-            )
-        }
-    }
-
-    func attachingPane(
-        _ newPaneID: String,
-        direction: ShellSplitDirection,
-        splitNodeID: String,
-        newLeafNodeID: String
-    ) -> ShellPaneTreeNode {
-        let newLeaf = ShellPaneTreeNode(
-            nodeID: newLeafNodeID,
-            kind: .pane,
-            direction: nil,
-            paneID: newPaneID,
-            children: nil
-        )
-
-        if kind == .split,
-           self.direction == direction,
-           let existingChildren = children,
-           let lastChild = existingChildren.last {
-            let nestedSplit = ShellPaneTreeNode(
-                nodeID: splitNodeID,
-                kind: .split,
-                direction: direction,
-                ratio: 0.5,
-                paneID: nil,
-                children: [lastChild, newLeaf]
-            )
-
-            return ShellPaneTreeNode(
-                nodeID: nodeID,
-                kind: .split,
-                direction: direction,
-                ratio: ratio,
-                paneID: nil,
-                children: Array(existingChildren.dropLast()) + [nestedSplit]
-            )
-        }
-
-        return ShellPaneTreeNode(
-            nodeID: splitNodeID,
-            kind: .split,
-            direction: direction,
-            ratio: 0.5,
-            paneID: nil,
-            children: [self, newLeaf]
-        )
-    }
-
-}
-
-enum ShellStateMutationError: String, Error {
-    case spaceNotFound = "space_not_found"
-    case tabNotFound = "tab_not_found"
-    case paneNotFound = "pane_not_found"
-    case splitNotFound = "split_not_found"
-    case spatialFocusTargetNotFound = "spatial_focus_target_not_found"
-    case lastTab = "last_tab"
-    case lastPane = "last_pane"
-    case invalidMoveTarget = "invalid_move_target"
-}
-
-struct ShellStateMutationResult {
-    let state: ShellStateSnapshot
-    let spaceID: String?
-    let tabID: String?
-    let paneID: String?
-}
-
-extension ShellStateSnapshot {
-    private static func defaultShellWorkingDirectory() -> String {
-        FileManager.default.homeDirectoryForCurrentUser.path
-    }
-
-    static func bootstrapDefault(
-        windowID: String = "window_main",
-        workingDirectory: String = defaultShellWorkingDirectory()
-    ) -> ShellStateSnapshot {
-        let spaceID = "space_main"
-        let tabID = "tab_main"
-        let paneID = "pane_1"
-
-        return ShellStateSnapshot(
-            contractVersion: "0.1",
-            windowID: windowID,
-            focusedSpaceID: spaceID,
-            focusedTabID: tabID,
-            focusedPaneID: paneID,
-            spaces: [
-                ShellSpace(
-                    spaceID: spaceID,
-                    title: "Terminal",
-                    attention: .active,
-                    tabs: [
-                        ShellTab(
-                            tabID: tabID,
-                            kind: .terminal,
-                            title: "Shell",
-                            paneTree: ShellPaneTreeNode(
-                                nodeID: "node_\(paneID)",
-                                kind: .pane,
-                                direction: nil,
-                                paneID: paneID,
-                                children: nil
-                            )
-                        )
-                    ]
-                )
-            ],
-            panes: [
-                ShellPane(
-                    paneID: paneID,
-                    tabID: tabID,
-                    spaceID: spaceID,
-                    launchTarget: .shell,
-                    cwd: workingDirectory,
-                    process: Self.defaultProcessBinding(for: .shell),
-                    attention: .active,
-                    context: nil,
-                    viewport: ShellViewportSnapshot(
-                        title: Self.defaultViewportTitle(for: .shell),
-                        summary: "ready to launch login shell",
-                        visibleExcerpt: nil,
-                        lastActivityAt: nil
-                    ),
-                    alanBinding: nil
-                )
-            ]
-        )
-    }
-
-
-    func focusingPane(_ paneID: String) throws -> ShellStateMutationResult {
-        guard pane(paneID: paneID) != nil else {
-            throw ShellStateMutationError.paneNotFound
-        }
-
-        return ShellStateMutationResult(
-            state: replacing(
-                spaces: spaces,
-                panes: panes,
-                focusedPaneID: paneID
-            ),
-            spaceID: pane(paneID: paneID)?.spaceID,
-            tabID: pane(paneID: paneID)?.tabID,
-            paneID: paneID
-        )
-    }
-
-    func focusingAdjacentPane(_ direction: ShellSpatialFocusDirection) throws -> ShellStateMutationResult {
-        guard let focusedPaneID,
-              let focusedPane = pane(paneID: focusedPaneID),
-              let focusedTab = tab(tabID: focusedPane.tabID)
-        else {
-            throw ShellStateMutationError.paneNotFound
-        }
-
-        guard let targetPaneID = focusedTab.paneTree.adjacentPaneID(
-            from: focusedPaneID,
-            direction: direction
-        ) else {
-            throw ShellStateMutationError.spatialFocusTargetNotFound
-        }
-
-        return try focusingPane(targetPaneID)
-    }
-
-    func creatingSpace(
-        launchTarget: ShellLaunchTarget,
-        title: String?,
-        workingDirectory: String?,
-        defaultWorkingDirectory: String = defaultShellWorkingDirectory(),
-        now: Date = .now
-    ) -> ShellStateMutationResult {
-        let spaceIndex = spaces.count + 1
-        let spaceID = nextID(prefix: "space", existing: spaces.map(\.spaceID))
-        let tabID = nextID(prefix: "tab", existing: spaces.flatMap { $0.tabs.map(\.tabID) })
-        let paneID = nextID(prefix: "pane", existing: panes.map(\.paneID))
-        let pane = makeTerminalPane(
-            paneID: paneID,
-            tabID: tabID,
-            spaceID: spaceID,
-            launchTarget: launchTarget,
-            workingDirectory: workingDirectory ?? defaultWorkingDirectory,
-            summary: launchTarget == .alan ? "new Alan space scaffolded" : "new shell space scaffolded",
-            now: now
-        )
-        let tab = ShellTab(
-            tabID: tabID,
-            kind: .terminal,
-            title: launchTarget == .alan ? "Alan" : "Shell",
-            paneTree: ShellPaneTreeNode(
-                nodeID: "node_\(paneID)",
-                kind: .pane,
-                direction: nil,
-                paneID: paneID,
-                children: nil
-            )
-        )
-        let space = ShellSpace(
-            spaceID: spaceID,
-            title: title ?? (launchTarget == .alan ? "Alan Space \(spaceIndex)" : "Space \(spaceIndex)"),
-            attention: .active,
-            tabs: [tab]
-        )
-        let nextPanes = panes + [pane]
-        let nextSpaces = rebuildingAttention(in: spaces + [space], panes: nextPanes)
-
-        return ShellStateMutationResult(
-            state: replacing(
-                spaces: nextSpaces,
-                panes: nextPanes,
-                focusedPaneID: paneID
-            ),
-            spaceID: spaceID,
-            tabID: tabID,
-            paneID: paneID
-        )
-    }
-
-    func creatingAlanSpace(
-        title: String?,
-        workingDirectory: String?,
-        defaultWorkingDirectory: String = defaultShellWorkingDirectory(),
-        now: Date = .now
-    ) -> ShellStateMutationResult {
-        creatingSpace(
-            launchTarget: .alan,
-            title: title,
-            workingDirectory: workingDirectory,
-            defaultWorkingDirectory: defaultWorkingDirectory,
-            now: now
-        )
-    }
-
-    func creatingTerminalSpace(
-        title: String?,
-        workingDirectory: String?,
-        defaultWorkingDirectory: String = defaultShellWorkingDirectory(),
-        now: Date = .now
-    ) -> ShellStateMutationResult {
-        creatingSpace(
-            launchTarget: .shell,
-            title: title,
-            workingDirectory: workingDirectory,
-            defaultWorkingDirectory: defaultWorkingDirectory,
-            now: now
-        )
-    }
-
-    func openingTab(
-        launchTarget: ShellLaunchTarget,
-        in requestedSpaceID: String?,
-        title: String?,
-        workingDirectory: String?,
-        defaultWorkingDirectory: String = defaultShellWorkingDirectory(),
-        now: Date = .now
-    ) throws -> ShellStateMutationResult {
-        let targetSpaceID = requestedSpaceID ?? focusedSpaceID ?? spaces.first?.spaceID
-        guard let targetSpaceID,
-              let targetSpace = space(spaceID: targetSpaceID)
-        else {
-            throw ShellStateMutationError.spaceNotFound
-        }
-
-        let tabID = nextID(prefix: "tab", existing: spaces.flatMap { $0.tabs.map(\.tabID) })
-        let paneID = nextID(prefix: "pane", existing: panes.map(\.paneID))
-        let pane = makeTerminalPane(
-            paneID: paneID,
-            tabID: tabID,
-            spaceID: targetSpaceID,
-            launchTarget: launchTarget,
-            workingDirectory: workingDirectory ?? defaultWorkingDirectory,
-            summary: launchTarget == .alan ? "new Alan tab scaffolded" : "new shell tab scaffolded",
-            now: now
-        )
-        let tab = ShellTab(
-            tabID: tabID,
-            kind: .terminal,
-            title: title ?? (launchTarget == .alan ? "Alan \(targetSpace.tabs.count + 1)" : "Shell \(targetSpace.tabs.count + 1)"),
-            paneTree: ShellPaneTreeNode(
-                nodeID: "node_\(paneID)",
-                kind: .pane,
-                direction: nil,
-                paneID: paneID,
-                children: nil
-            )
-        )
-        let nextSpaces = spaces.map { space in
-            guard space.spaceID == targetSpaceID else { return space }
-            return ShellSpace(
-                spaceID: space.spaceID,
-                title: space.title,
-                attention: space.attention,
-                tabs: space.tabs + [tab]
-            )
-        }
-        let nextPanes = panes + [pane]
-
-        return ShellStateMutationResult(
-            state: replacing(
-                spaces: rebuildingAttention(in: nextSpaces, panes: nextPanes),
-                panes: nextPanes,
-                focusedPaneID: paneID
-            ),
-            spaceID: targetSpaceID,
-            tabID: tabID,
-            paneID: paneID
-        )
-    }
-
-    func openingAlanTab(
-        in requestedSpaceID: String?,
-        title: String?,
-        workingDirectory: String?,
-        defaultWorkingDirectory: String = defaultShellWorkingDirectory(),
-        now: Date = .now
-    ) throws -> ShellStateMutationResult {
-        try openingTab(
-            launchTarget: .alan,
-            in: requestedSpaceID,
-            title: title,
-            workingDirectory: workingDirectory,
-            defaultWorkingDirectory: defaultWorkingDirectory,
-            now: now
-        )
-    }
-
-    func openingTerminalTab(
-        in requestedSpaceID: String?,
-        title: String?,
-        workingDirectory: String?,
-        defaultWorkingDirectory: String = defaultShellWorkingDirectory(),
-        now: Date = .now
-    ) throws -> ShellStateMutationResult {
-        try openingTab(
-            launchTarget: .shell,
-            in: requestedSpaceID,
-            title: title,
-            workingDirectory: workingDirectory,
-            defaultWorkingDirectory: defaultWorkingDirectory,
-            now: now
-        )
-    }
-
-    func splittingPane(
-        _ paneID: String,
-        direction: ShellSplitDirection,
-        defaultWorkingDirectory: String = defaultShellWorkingDirectory(),
-        now: Date = .now
-    ) throws -> ShellStateMutationResult {
-        try splittingPane(
-            paneID,
-            placement: .defaultPlacement(for: direction),
-            defaultWorkingDirectory: defaultWorkingDirectory,
-            now: now
-        )
-    }
-
-    func splittingPane(
-        _ paneID: String,
-        placement: ShellPaneSplitDirection,
-        defaultWorkingDirectory: String = defaultShellWorkingDirectory(),
-        now: Date = .now
-    ) throws -> ShellStateMutationResult {
-        guard let pane = pane(paneID: paneID),
-              let tab = tab(tabID: pane.tabID)
-        else {
-            throw ShellStateMutationError.paneNotFound
-        }
-
-        let newPaneID = nextID(prefix: "pane", existing: panes.map(\.paneID))
-        let splitNodeID = nextID(
-            prefix: "node",
-            existing: spaces.flatMap(\.tabs).flatMap { $0.paneTree.nodeIDs }
-        )
-        let newPane = makeTerminalPane(
-            paneID: newPaneID,
-            tabID: pane.tabID,
-            spaceID: pane.spaceID,
-            launchTarget: pane.resolvedLaunchTarget,
-            workingDirectory: pane.cwd ?? defaultWorkingDirectory,
-            summary: "new split scaffolded",
-            now: now
-        )
-        let updatedTab = ShellTab(
-            tabID: tab.tabID,
-            kind: tab.kind,
-            title: tab.title,
-            paneTree: tab.paneTree.splittingPane(
-                paneID,
-                placement: placement,
-                splitNodeID: splitNodeID,
-                newLeafNodeID: "node_\(newPaneID)",
-                newPaneID: newPaneID
-            )
-        )
-
-        let nextSpaces = spaces.map { space in
-            guard space.spaceID == pane.spaceID else { return space }
-            return ShellSpace(
-                spaceID: space.spaceID,
-                title: space.title,
-                attention: space.attention,
-                tabs: space.tabs.map { existingTab in
-                    existingTab.tabID == updatedTab.tabID ? updatedTab : existingTab
-                }
-            )
-        }
-        let nextPanes = panes + [newPane]
-
-        return ShellStateMutationResult(
-            state: replacing(
-                spaces: rebuildingAttention(in: nextSpaces, panes: nextPanes),
-                panes: nextPanes,
-                focusedPaneID: newPaneID
-            ),
-            spaceID: pane.spaceID,
-            tabID: pane.tabID,
-            paneID: newPaneID
-        )
-    }
-
-    func resizingSplit(
-        _ splitNodeID: String,
-        ratio: Double
-    ) throws -> ShellStateMutationResult {
-        guard let tab = spaces.lazy
-            .flatMap(\.tabs)
-            .first(where: { $0.paneTree.contains(nodeID: splitNodeID) })
-        else {
-            throw ShellStateMutationError.splitNotFound
-        }
-
-        let resizeResult = tab.paneTree.resizingSplit(splitNodeID, ratio: ratio)
-        guard resizeResult.changed else {
-            throw ShellStateMutationError.splitNotFound
-        }
-
-        return try replacingTabTree(
-            tabID: tab.tabID,
-            paneTree: resizeResult.node
-        )
-    }
-
-    func equalizingSplits(in requestedTabID: String?) throws -> ShellStateMutationResult {
-        let tabID = requestedTabID ?? focusedTabID
-        guard let tabID,
-              let tab = tab(tabID: tabID)
-        else {
-            throw ShellStateMutationError.tabNotFound
-        }
-
-        return try replacingTabTree(
-            tabID: tab.tabID,
-            paneTree: tab.paneTree.equalizedSplits()
-        )
-    }
-
-    func closingPane(_ paneID: String) throws -> ShellStateMutationResult {
-        guard let pane = pane(paneID: paneID),
-              let tab = tab(tabID: pane.tabID)
-        else {
-            throw ShellStateMutationError.paneNotFound
-        }
-
-        if tab.paneTree.paneIDs.count == 1 {
-            return try closingTab(tab.tabID)
-        }
-
-        guard let updatedPaneTree = tab.paneTree.removingPane(paneID) else {
-            throw ShellStateMutationError.paneNotFound
-        }
-
-        let updatedTab = ShellTab(
-            tabID: tab.tabID,
-            kind: tab.kind,
-            title: tab.title,
-            paneTree: updatedPaneTree
-        )
-        let nextSpaces = spaces.map { space in
-            guard space.spaceID == pane.spaceID else { return space }
-            return ShellSpace(
-                spaceID: space.spaceID,
-                title: space.title,
-                attention: space.attention,
-                tabs: space.tabs.map { existingTab in
-                    existingTab.tabID == updatedTab.tabID ? updatedTab : existingTab
-                }
-            )
-        }
-        let nextPanes = panes.filter { $0.paneID != paneID }
-        let preferredPaneID =
-            focusedPaneID == paneID
-            ? (updatedTab.paneTree.paneIDs.first
-                ?? nextPanes.first(where: { $0.spaceID == pane.spaceID })?.paneID
-                ?? nextPanes.first?.paneID)
-            : focusedPaneID
-
-        let nextState = replacing(
-            spaces: rebuildingAttention(in: nextSpaces, panes: nextPanes),
-            panes: nextPanes,
-            focusedPaneID: preferredPaneID
-        )
-
-        return ShellStateMutationResult(
-            state: nextState,
-            spaceID: pane.spaceID,
-            tabID: pane.tabID,
-            paneID: nextState.focusedPaneID
-        )
-    }
-
-    func movingPaneToNewTab(
-        _ paneID: String,
-        title: String?,
-        now: Date = .now
-    ) throws -> ShellStateMutationResult {
-        guard let pane = pane(paneID: paneID),
-              let sourceTab = tab(tabID: pane.tabID)
-        else {
-            throw ShellStateMutationError.paneNotFound
-        }
-
-        guard sourceTab.paneTree.paneIDs.count > 1 else {
-            throw ShellStateMutationError.lastPane
-        }
-
-        guard let sourcePaneTree = sourceTab.paneTree.removingPane(paneID) else {
-            throw ShellStateMutationError.paneNotFound
-        }
-
-        let newTabID = nextID(
-            prefix: "tab",
-            existing: spaces.flatMap { $0.tabs.map(\.tabID) }
-        )
-        let movedPaneNode = ShellPaneTreeNode(
-            nodeID: "node_\(paneID)",
-            kind: .pane,
-            direction: nil,
-            paneID: paneID,
-            children: nil
-        )
-        let updatedSourceTab = ShellTab(
-            tabID: sourceTab.tabID,
-            kind: sourceTab.kind,
-            title: sourceTab.title,
-            paneTree: sourcePaneTree
-        )
-        let newTab = ShellTab(
-            tabID: newTabID,
-            kind: sourceTab.kind,
-            title: title ?? pane.viewport?.title ?? "Lifted Pane",
-            paneTree: movedPaneNode
-        )
-
-        let nextSpaces = spaces.map { space in
-            guard space.spaceID == pane.spaceID else { return space }
-            return ShellSpace(
-                spaceID: space.spaceID,
-                title: space.title,
-                attention: space.attention,
-                tabs: space.tabs.flatMap { existingTab -> [ShellTab] in
-                    guard existingTab.tabID == sourceTab.tabID else {
-                        return [existingTab]
-                    }
-                    return [updatedSourceTab, newTab]
-                }
-            )
-        }
-        let formatter = ISO8601DateFormatter()
-        let nextPanes = panes.map { current in
-            guard current.paneID == paneID else { return current }
-            return ShellPane(
-                paneID: current.paneID,
-                tabID: newTabID,
-                spaceID: current.spaceID,
-                launchTarget: current.launchTarget,
-                cwd: current.cwd,
-                process: current.process,
-                attention: current.attention,
-                context: current.context,
-                viewport: ShellViewportSnapshot(
-                    title: current.viewport?.title,
-                    summary: current.viewport?.summary ?? "pane moved to its own tab",
-                    visibleExcerpt: current.viewport?.visibleExcerpt,
-                    lastActivityAt: formatter.string(from: now)
-                ),
-                alanBinding: current.alanBinding
-            )
-        }
-
-        return ShellStateMutationResult(
-            state: replacing(
-                spaces: rebuildingAttention(in: nextSpaces, panes: nextPanes),
-                panes: nextPanes,
-                focusedPaneID: paneID
-            ),
-            spaceID: pane.spaceID,
-            tabID: newTabID,
-            paneID: paneID
-        )
-    }
-
-    func movingPane(
-        _ paneID: String,
-        toTab targetTabID: String,
-        direction: ShellSplitDirection,
-        now: Date = .now
-    ) throws -> ShellStateMutationResult {
-        guard let pane = pane(paneID: paneID),
-              let sourceTab = tab(tabID: pane.tabID)
-        else {
-            throw ShellStateMutationError.paneNotFound
-        }
-
-        guard let targetTab = tab(tabID: targetTabID) else {
-            throw ShellStateMutationError.tabNotFound
-        }
-
-        guard sourceTab.tabID != targetTab.tabID else {
-            throw ShellStateMutationError.invalidMoveTarget
-        }
-
-        let targetSpaceID = spaces.first(where: { space in
-            space.tabs.contains(where: { $0.tabID == targetTabID })
-        })?.spaceID
-
-        guard let targetSpaceID else {
-            throw ShellStateMutationError.tabNotFound
-        }
-
-        let formatter = ISO8601DateFormatter()
-        let moveSummary = "pane moved to \(targetTab.title ?? targetTab.tabID)"
-        let newSplitNodeID = nextID(
-            prefix: "node",
-            existing: spaces.flatMap { $0.tabs.flatMap { $0.paneTree.nodeIDs } }
-        )
-        let newLeafNodeID = "node_\(paneID)_moved"
-
-        let updatedTargetTab = ShellTab(
-            tabID: targetTab.tabID,
-            kind: targetTab.kind,
-            title: targetTab.title,
-            paneTree: targetTab.paneTree.attachingPane(
-                paneID,
-                direction: direction,
-                splitNodeID: newSplitNodeID,
-                newLeafNodeID: newLeafNodeID
-            )
-        )
-
-        let updatedSourcePaneTree = sourceTab.paneTree.removingPane(paneID)
-
-        let nextSpaces = spaces.compactMap { space -> ShellSpace? in
-            var nextTabs: [ShellTab] = []
-            for tab in space.tabs {
-                if tab.tabID == sourceTab.tabID {
-                    if let updatedSourcePaneTree {
-                        nextTabs.append(
-                            ShellTab(
-                                tabID: sourceTab.tabID,
-                                kind: sourceTab.kind,
-                                title: sourceTab.title,
-                                paneTree: updatedSourcePaneTree
-                            )
-                        )
-                    }
-                    continue
-                }
-
-                if tab.tabID == updatedTargetTab.tabID {
-                    nextTabs.append(updatedTargetTab)
-                } else {
-                    nextTabs.append(tab)
-                }
-            }
-
-            guard !nextTabs.isEmpty else { return nil }
-            return ShellSpace(
-                spaceID: space.spaceID,
-                title: space.title,
-                attention: space.attention,
-                tabs: nextTabs
-            )
-        }
-
-        let nextPanes = panes.map { current in
-            guard current.paneID == paneID else { return current }
-            return ShellPane(
-                paneID: current.paneID,
-                tabID: updatedTargetTab.tabID,
-                spaceID: targetSpaceID,
-                launchTarget: current.launchTarget,
-                cwd: current.cwd,
-                process: current.process,
-                attention: current.attention,
-                context: current.context,
-                viewport: ShellViewportSnapshot(
-                    title: current.viewport?.title,
-                    summary: current.viewport?.summary ?? moveSummary,
-                    visibleExcerpt: current.viewport?.visibleExcerpt,
-                    lastActivityAt: formatter.string(from: now)
-                ),
-                alanBinding: current.alanBinding
-            )
-        }
-
-        let nextState = replacing(
-            spaces: rebuildingAttention(in: nextSpaces, panes: nextPanes),
-            panes: nextPanes,
-            focusedPaneID: paneID
-        )
-
-        return ShellStateMutationResult(
-            state: nextState,
-            spaceID: targetSpaceID,
-            tabID: updatedTargetTab.tabID,
-            paneID: paneID
-        )
-    }
-
-    func settingAttention(
-        _ attention: ShellAttentionState,
-        for paneID: String
-    ) throws -> ShellStateMutationResult {
-        guard pane(paneID: paneID) != nil else {
-            throw ShellStateMutationError.paneNotFound
-        }
-        let nextPanes = panes.map { current in
-            guard current.paneID == paneID else { return current }
-            return ShellPane(
-                paneID: current.paneID,
-                tabID: current.tabID,
-                spaceID: current.spaceID,
-                launchTarget: current.launchTarget,
-                cwd: current.cwd,
-                process: current.process,
-                attention: attention,
-                context: current.context,
-                viewport: current.viewport,
-                alanBinding: current.alanBinding
-            )
-        }
-
-        return ShellStateMutationResult(
-            state: replacing(
-                spaces: rebuildingAttention(in: spaces, panes: nextPanes),
-                panes: nextPanes,
-                focusedPaneID: focusedPaneID ?? paneID
-            ),
-            spaceID: pane(paneID: paneID)?.spaceID,
-            tabID: pane(paneID: paneID)?.tabID,
-            paneID: paneID
-        )
-    }
-
-    private func replacingTabTree(
-        tabID: String,
-        paneTree: ShellPaneTreeNode
-    ) throws -> ShellStateMutationResult {
-        guard let targetSpace = spaces.first(where: { space in
-            space.tabs.contains(where: { $0.tabID == tabID })
-        }),
-        let targetTab = targetSpace.tabs.first(where: { $0.tabID == tabID })
-        else {
-            throw ShellStateMutationError.tabNotFound
-        }
-
-        let updatedTab = ShellTab(
-            tabID: targetTab.tabID,
-            kind: targetTab.kind,
-            title: targetTab.title,
-            paneTree: paneTree
-        )
-        let nextSpaces = spaces.map { space in
-            guard space.spaceID == targetSpace.spaceID else { return space }
-            return ShellSpace(
-                spaceID: space.spaceID,
-                title: space.title,
-                attention: space.attention,
-                tabs: space.tabs.map { tab in
-                    tab.tabID == updatedTab.tabID ? updatedTab : tab
-                }
-            )
-        }
-
-        return ShellStateMutationResult(
-            state: replacing(
-                spaces: nextSpaces,
-                panes: panes,
-                focusedPaneID: focusedPaneID
-            ),
-            spaceID: targetSpace.spaceID,
-            tabID: updatedTab.tabID,
-            paneID: focusedPaneID
-        )
-    }
-
-    func closingTab(_ tabID: String) throws -> ShellStateMutationResult {
-        guard totalTabCount > 1 else {
-            throw ShellStateMutationError.lastTab
-        }
-        guard let targetSpace = spaces.first(where: { space in
-            space.tabs.contains(where: { $0.tabID == tabID })
-        }),
-        let targetTab = targetSpace.tabs.first(where: { $0.tabID == tabID })
-        else {
-            throw ShellStateMutationError.tabNotFound
-        }
-
-        let removedPaneIDs = Set(targetTab.paneTree.paneIDs)
-        let nextSpaces = spaces.compactMap { space -> ShellSpace? in
-            let remainingTabs = space.tabs.filter { $0.tabID != tabID }
-            guard !remainingTabs.isEmpty else {
-                return space.spaceID == targetSpace.spaceID ? nil : space
-            }
-            return ShellSpace(
-                spaceID: space.spaceID,
-                title: space.title,
-                attention: space.attention,
-                tabs: remainingTabs
-            )
-        }
-        let nextPanes = panes.filter { !removedPaneIDs.contains($0.paneID) }
-        let preferredPaneID = nextPanes.first(where: { $0.spaceID == targetSpace.spaceID })?.paneID
-            ?? nextPanes.first?.paneID
-        let nextState = replacing(
-            spaces: rebuildingAttention(in: nextSpaces, panes: nextPanes),
-            panes: nextPanes,
-            focusedPaneID: preferredPaneID
-        )
-
-        return ShellStateMutationResult(
-            state: nextState,
-            spaceID: nextState.focusedSpaceID,
-            tabID: tabID,
-            paneID: nextState.focusedPaneID
-        )
-    }
-
-    private func replacing(
-        spaces: [ShellSpace],
-        panes: [ShellPane],
-        focusedPaneID: String?
-    ) -> ShellStateSnapshot {
-        let resolvedFocusedPaneID =
-            focusedPaneID.flatMap { candidate in
-                panes.contains(where: { $0.paneID == candidate }) ? candidate : nil
-            } ?? panes.first?.paneID
-        let focusedPane = resolvedFocusedPaneID.flatMap { candidate in
-            panes.first(where: { $0.paneID == candidate })
-        }
-
-        return ShellStateSnapshot(
-            contractVersion: contractVersion,
-            windowID: windowID,
-            focusedSpaceID: focusedPane?.spaceID ?? spaces.first?.spaceID,
-            focusedTabID: focusedPane?.tabID ?? spaces.first?.tabs.first?.tabID,
-            focusedPaneID: resolvedFocusedPaneID,
-            spaces: spaces,
-            panes: panes
-        )
-    }
-
-    private func rebuildingAttention(in spaces: [ShellSpace], panes: [ShellPane]) -> [ShellSpace] {
-        spaces.map { space in
-            ShellSpace(
-                spaceID: space.spaceID,
-                title: space.title,
-                attention: strongestAttention(in: panes.filter { $0.spaceID == space.spaceID }),
-                tabs: space.tabs
-            )
-        }
-    }
-
-    private func strongestAttention(in panes: [ShellPane]) -> ShellAttentionState {
-        panes
-            .map(\.attention)
-            .max(by: { Self.attentionRank(for: $0) < Self.attentionRank(for: $1) })
-            ?? .idle
-    }
-
-    private static func attentionRank(for attention: ShellAttentionState) -> Int {
-        switch attention {
-        case .idle:
-            return 0
-        case .active:
-            return 1
-        case .notable:
-            return 2
-        case .awaitingUser:
-            return 3
-        }
-    }
-
-    private func nextID(prefix: String, existing: [String]) -> String {
-        let nextOrdinal = existing
-            .compactMap { identifier -> Int? in
-                let components = identifier.split(separator: "_")
-                guard let last = components.last else { return nil }
-                return Int(last)
-            }
-            .max()
-            .map { $0 + 1 }
-            ?? (existing.isEmpty ? 1 : existing.count + 1)
-
-        return "\(prefix)_\(nextOrdinal)"
-    }
-
-    private func makeTerminalPane(
-        paneID: String,
-        tabID: String,
-        spaceID: String,
-        launchTarget: ShellLaunchTarget,
-        workingDirectory: String,
-        summary: String,
-        now: Date
-    ) -> ShellPane {
-        let formatter = ISO8601DateFormatter()
-        return ShellPane(
-            paneID: paneID,
-            tabID: tabID,
-            spaceID: spaceID,
-            launchTarget: launchTarget,
-            cwd: workingDirectory,
-            process: Self.defaultProcessBinding(for: launchTarget),
-            attention: .active,
-            context: nil,
-            viewport: ShellViewportSnapshot(
-                title: Self.defaultViewportTitle(for: launchTarget),
-                summary: summary,
-                visibleExcerpt: nil,
-                lastActivityAt: formatter.string(from: now)
-            ),
-            alanBinding: nil
-        )
-    }
-
-    private static func defaultProcessBinding(for launchTarget: ShellLaunchTarget) -> ShellProcessBinding {
-        switch launchTarget {
-        case .shell:
-            let shellPath = defaultShellPath()
-            return ShellProcessBinding(
-                program: URL(fileURLWithPath: shellPath).lastPathComponent.isEmpty
-                    ? "zsh"
-                    : URL(fileURLWithPath: shellPath).lastPathComponent,
-                argvPreview: ["-l"]
-            )
-        case .alan:
-            return ShellProcessBinding(program: "alan-tui", argvPreview: ["alan", "chat"])
-        }
-    }
-
-    private static func defaultViewportTitle(for launchTarget: ShellLaunchTarget) -> String {
-        switch launchTarget {
-        case .shell:
-            return "Shell"
-        case .alan:
-            return "Alan"
-        }
-    }
-
-    private static func defaultShellPath(
-        environment: [String: String] = ProcessInfo.processInfo.environment
-    ) -> String {
-        let shell = environment["SHELL"]?.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard let shell, !shell.isEmpty else {
-            return "/bin/zsh"
-        }
-        return shell
-    }
-}
-
-extension ShellStateSnapshot {
-    static let spikePreview = ShellStateSnapshot(
-        contractVersion: "0.1",
-        windowID: "window_main",
-        focusedSpaceID: "space_alan_app",
-        focusedTabID: "tab_main",
-        focusedPaneID: "pane_1",
-        spaces: [
-            ShellSpace(
-                spaceID: "space_alan_app",
-                title: "Alan App",
-                attention: .awaitingUser,
-                tabs: [
-                    ShellTab(
-                        tabID: "tab_main",
-                        kind: .terminal,
-                        title: "Main Session",
-                        paneTree: ShellPaneTreeNode(
-                            nodeID: "node_root",
-                            kind: .split,
-                            direction: .vertical,
-                            paneID: nil,
-                            children: [
-                                ShellPaneTreeNode(
-                                    nodeID: "pane_1",
-                                    kind: .pane,
-                                    direction: nil,
-                                    paneID: "pane_1",
-                                    children: nil
-                                ),
-                                ShellPaneTreeNode(
-                                    nodeID: "pane_2",
-                                    kind: .pane,
-                                    direction: nil,
-                                    paneID: "pane_2",
-                                    children: nil
-                                ),
-                            ]
-                        )
-                    )
-                ]
-            )
-        ],
-        panes: [
-            ShellPane(
-                paneID: "pane_1",
-                tabID: "tab_main",
-                spaceID: "space_alan_app",
-                launchTarget: .alan,
-                cwd: "/Users/morris/Developer/Alan",
-                process: ShellProcessBinding(program: "alan-tui", argvPreview: ["alan", "chat"]),
-                attention: .awaitingUser,
-                context: ShellContextSnapshot(
-                    workingDirectoryName: "Alan",
-                    repositoryRoot: "/Users/morris/Developer/Alan",
-                    gitBranch: "main",
-                    controlPath: "/tmp/alan-shell-control/window_main",
-                    alanBindingFile: "/tmp/alan-shell-control/window_main/panes/pane_1/alan-binding.json",
-                    launchStrategy: "installed_binary",
-                    shellIntegrationSource: "ghostty_shell_integration",
-                    processState: "running",
-                    lastMetadataAt: "2026-04-01T10:30:00Z",
-                    lastCommandExitCode: nil
-                ),
-                viewport: ShellViewportSnapshot(
-                    title: "Alan",
-                    summary: "waiting for approval",
-                    visibleExcerpt: nil,
-                    lastActivityAt: "2026-04-01T10:30:00Z"
-                ),
-                alanBinding: ShellAlanBinding(
-                    sessionID: "sess_123",
-                    runStatus: "yielded",
-                    pendingYield: true,
-                    source: "preview",
-                    lastProjectedAt: "2026-04-01T10:30:00Z"
-                )
-            ),
-            ShellPane(
-                paneID: "pane_2",
-                tabID: "tab_main",
-                spaceID: "space_alan_app",
-                launchTarget: .shell,
-                cwd: "/Users/morris/Developer/Alan",
-                process: ShellProcessBinding(program: "zsh", argvPreview: nil),
-                attention: .idle,
-                context: ShellContextSnapshot(
-                    workingDirectoryName: "Alan",
-                    repositoryRoot: "/Users/morris/Developer/Alan",
-                    gitBranch: "main",
-                    controlPath: "/tmp/alan-shell-control/window_main",
-                    alanBindingFile: "/tmp/alan-shell-control/window_main/panes/pane_2/alan-binding.json",
-                    launchStrategy: "path_binary",
-                    shellIntegrationSource: "ghostty_shell_integration",
-                    processState: "running",
-                    lastMetadataAt: "2026-04-01T10:24:00Z",
-                    lastCommandExitCode: 0
-                ),
-                viewport: ShellViewportSnapshot(
-                    title: "shell",
-                    summary: "idle shell",
-                    visibleExcerpt: nil,
-                    lastActivityAt: nil
-                ),
-                alanBinding: nil
-            ),
-        ]
-    )
-}
